ensure 64bit type for bitfield
size_t is not guaranteed to be >= 64 bit in length. This fixes
building minimap on architectures where it's generally less (i.e.
32-bit platforms).

+++ kdq.h
@@ -1,12 +1,13 @@
 #ifndef __AC_KDQ_H
 #define __AC_KDQ_H
 
+#include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 
 #define __KDQ_TYPE(type) \
 	typedef struct { \
-		size_t front:58, bits:6, count, mask; \
+		uint64_t front:58, bits:6, count, mask; \
 		type *a; \
 	} kdq_##type##_t;
